making it work on primary
